Implement a test for constructing a Round object
This change implements a test that verifies we have written the method
to create a Round object from each line of an input file with opponent
and player moves in one round of Rock-Paper-Scissors. We define
enumerations for the valid move options for each player per the format
the challenge specifies. I.e., "A", "B", and "C" for Rock, Paper, and
Scissors from the opponent and "X", "Y", and "Z" for Rock, Paper, and
Scissors for the player.