Adapt the code for Part 2 of Day 2
This change adapts our code to handle the new information Part 2 of the
Day 2 challenge presents. Specifically, we learn that the second column
of data does not describe the move the player makes, but, instead, the
outcome the player should create. We enumerate those outcome values in
a class in day_02.py. We create a method in the Round class to determine
the move the player should make given a desired outcome and the move the
opponent makes. Finally, we create a flag we pass to the from_file
method of Game to determine which interpretation to use when it
creates Round objects from the lines of the input file.